I added past presentations and more types + stdlib operators to coloring.
I made a folder containing all of my past presentations made for \English.
This is just there for anyone interested.

I also added some more types and standard operators to the custom Gedit
coloring scheme, including the "regex" type. I also added a context that
matches "regexp" and marks the "p" as an error because that "p" is a pet
peeve of mine. :)

Finally, I added a couple examples of operator definitions commented for
understanding to the examples folder.

Signed-off-by: REALDrummer <conman.davey@gmail.com>