I fixed the "read" operator in examples and coloring.
First, I found an issue with one of my examples in that it was trying to
convert a string to a byte rather than "read" the string to a byte. Since
these may be different as taking a string as a byte might be asking it to
grab the first char of the string as a byte or something equally weird,
I eliminated the ambiguity by using the "read" operator instead.

Second, this made me realize that the "read" operator was not properly
defined in the Gedit coloring scheme. I not only added the word "read" by
itself to the list of available operators in the "common defines", but I
also gave read its own context definition to color the "as a" of the
"read <string> as a <type>" properly.

Signed-off-by: REALDrummer <conman.davey@gmail.com>