I added a new Java port example and some big questions.
I had this little Java plugin for Minecraft that was supposed to just find
all the wolves on the server and print their locations to any ops on the
server, then print the location if any new wolves spawn. I took that Java
plugin and rewrote it in \English and included both as an example.

I also typed up some explanations of some of the big questions left in the
design of this language and posted them to the Trello board. There are now
copies of those questions here as well.

Signed-off-by: REALDrummer <conman.davey@gmail.com>