I added a bunch of compiler code and a new \English example.
To the compiler code, I added and reorganized many classes as I began to
start in on the system for parsing operator declarations. This system is
implemented in the constructors for `op_syntax` and its derived classes.

I also added some more messaging/debugging utilities relating to the use
of file locations in messages; see `LOCATIONAL_ERR`, `LOCATIONAL_WARN`, etc.

Finally, I added a quick sort example for \English code.

Signed-off-by: REALDrummer <conman.davey@gmail.com>