I made the outline for the usage documentation and tweaked some examples.
It came to my attention that a document detailing how to use \English
would be very useful, so I wrote up a LaTeX outline for it and will fill
it in with all sorts of info ASAP.

In the `quick sort` and `print staircase` examples, I decided to try out
the idea of delimiting blocks via indentation like Python with the little
difference that commas start blocks rather than colons. In addition, in
the quick sort, I tried replacing the pattern-matching `@` with a kind of
partially applied operator idea I had.

Signed-off-by: REALDrummer <conman.davey@gmail.com>