I added packages and commands and made a couple subsections.
As you'll probably note, I used a bunch of packages and custom commands and
things to configure the environment. Most of them have comments to explain
why they are there.

I also filled in the sections about comments and number types. There's a
lot more to go, but I've been kind of fighting with LaTeX to get this far.
The content isn't the issue; it's making LaTeX not screw up. Macros in
LaTeX are more buggy and side-effect-prone than C's text-replace
preprocessor.

On a related note, you'll notice that the first section is numbered '2'
and that at the the top of the table of contents, there's a '1' section
with a "*" instead of a name. I don't know how or why. I'm trying to fix it.

Signed-off-by: REALDrummer <conman.davey@gmail.com>