I filled in several sections and fixed some bugs.
I filled in stuff about declaring variables, using numbers and math, and
the start of using lists in \English. Check it out in the PDF!

I fixed that weird thing with that first "*" section. However, I seem to
have a new bug that's spitting out a bunch of random junk at the top of
the table of contents. I still have no idea how or why.

Signed-off-by: REALDrummer <conman.davey@gmail.com>