This is a large commit.
For one thing, this commit adds the parser class. One parser is meant to parse one file, keeping track of things like the current line number and character number for error outputs. This one-file-per-parser is not only simple and scalable, but also allows for the possibility of concurrent file parsing. This also replaced the messaging and debugging utilities header file.

I also added 

I also added the values class and some derived classes of it. This class is meant to represent one Assembly-level value, e.g. an Assembly label, a register, or a memory address.

I also added a new example.

Finally, I added a LaTeX outline for a possible style guide document to start.