I added the '//' operator and arbitrary base numbers to the usage docs.
The '//' operator is a binary integer division operator. The standard '/' will no longer be an integer division even if both of its arguments are integers unless the first operand is divisible by the second; see the usage docs for more info.

I also added the ability to put an underscore and an integer between 1 and 36 (inclusive) after any number for arbitrary base numbers.

Also, I changed the custom preprocess command in the LaTeX to codespecial because it is a more accurate description of what the command is used to highlight (though maybe more vague).